[KP-20] Splitted Builders into Modules
+ rename MainModule and DetailModule and add DispatchQueue on NetworkService's fetched data
