+++ KinoPoisk/Modules/DetailModule/Presenter/DetailTableViewPresenter.swift
